Rename from App* to The* to follow Vue conventions
